Update JUnit to 5.7.1 and avoid to approved stacktrace
Need to update JUnit to handle TempFile under windows for CI in Github

Some lines of stacktrace was displayed in an approved file.
When stacktrace change, the test failed but often because line number change between OS or version library.
Now, we write Stacktrace in a file and include it in final document.
Stacktrace still visible but was not anymore checked by approvals.